+++ docs/CHANGELOG.md
@@ -102,8 +102,8 @@ Semantic Versioning.
     variables without diagnostics.
   * Fixed false negatives for [E0196][] if a type has the same name as a
     variable.
-  * Using a variable in a type (e.g. with `typeof`) no longer falsely reports
-    [E0058][] ("variable used before declaration").
+  * Using a variable in a type or interface (e.g. with `typeof`) no longer
+    falsely reports [E0058][] ("variable used before declaration").
   * Type predicates in function types no longer falsely report [E0315][]
     ("'param' is not the name of a parameter").